feat : State 컬럼 재설정
